requirements.txt
tensorflow-cpu==2.17.0
numpy==1.26.4
uvicorn==0.32.0
fastapi==0.115.3
pydantic==2.9.2
starlette==0.41.0
pytest==8.3.3
httpx
opencensus-ext-azure==1.1.13
opencensus-ext-requests==0.8.0
azure-core==1.32.0
azure-identity==1.19.0
opencensus==0.11.4
opencensus-context==0.1.3
msal==1.31.1
msal-extensions==1.2.0
cryptography==43.0.3
portalocker==2.10.1
google-api-core==2.23.0
proto-plus==1.25.0
PyJWT==2.10.0
python-dotenv
azure-storage-blob==12.19.1
